+++ lib/representable/definition.rb
@@ -1,16 +1,35 @@
 module Representable
   # Created at class compile time. Keeps configuration options for one property.
-  class Definition
+  class Definition < Hash
     attr_reader :name, :options
     alias_method :getter, :name
 
     def initialize(sym, options={})
       @name     = sym.to_s
-      @options  = options
+
+      # deprecations:
+      raise "The :from option got replaced by :as in Representable 1.8!" if options[:from]
+
+      super()
+
+      self[:as]  = (options.delete(:as) || @name).to_s
+
+      options.each { |k,v| self[k] = v }
     end
 
+    private :merge!, :default
+
     def clone
-      self.class.new(name, options.clone) # DISCUSS: make generic Definition.cloned_attribute that passes list to constructor.
+      self.class.new(name, super) # DISCUSS: make generic Definition.cloned_attribute that passes list to constructor.
+    end
+
+    def as
+      self[:as]
+    end
+
+    def options
+      # TODO: deprecate.
+      self
     end
 
     def setter
@@ -18,65 +37,44 @@ module Representable
     end
 
     def typed?
-      deserialize_class.is_a?(Class) or representer_module or options[:instance]  # also true if only :extend is set, for people who want solely rendering.
+      deserialize_class.is_a?(Class) or representer_module or self[:instance]  # also true if only :extend is set, for people who want solely rendering.
     end
 
     def array?
-      options[:collection]
+      self[:collection]
     end
 
     def hash?
-      options[:hash]
+      self[:hash]
     end
 
     def deserialize_class
-      options[:class]
-    end
-
-    def from
-      # TODO: deprecate :from.
-      (options[:from] || options[:as] || name).to_s
+      self[:class]
     end
 
     def default_for(value)
-      return default if skipable_nil_value?(value)
+      return self[:default] if skipable_nil_value?(value)
       value
     end
 
     def has_default?
-      options.has_key?(:default)
+      has_key?(:default)
     end
 
     def representer_module
-      options[:extend] or options[:decorator]
-    end
-
-    def attribute
-      options[:attribute]
-    end
-
-    def content
-      options[:content]
+      self[:extend] or self[:decorator]
     end
 
     def skipable_nil_value?(value)
-      value.nil? and not options[:render_nil]
-    end
-
-    def default
-      options[:default]
-    end
-
-    def binding
-      options[:binding]
+      value.nil? and not self[:render_nil]
     end
 
     def create_binding(*args)
-      binding.call(self, *args)
+      self[:binding].call(self, *args)
     end
 
     def sync?
-      options[:parse_strategy] == :sync
+      self[:parse_strategy] == :sync
     end
   end
 end