+++ test/inherit_test.rb
@@ -67,10 +67,27 @@ class InheritTest < MiniTest::Spec
       include SongRepresenter
 
       property :name, :inherit => true, :as => :name
+      puts representable_attrs.inspect
     end
 
     it { SongRepresenter.prepare(Song.new(Struct.new(:string).new("Believe It"), 1)).to_hash.must_equal({"title"=>{"str"=>"Believe It"}, "no"=>1}) }
     it { representer.prepare( Song.new(Struct.new(:string).new("Believe It"), 1)).to_hash.must_equal({"name"=>{"str"=>"Believe It"}, "no"=>1}) }
+  end
+
+
+
+  # no :inherit
+  describe "overwriting without :inherit" do
+    representer! do
+      include SongRepresenter
+
+      property :track, :representable => true
+    end
+
+    it "replaces inherited property" do
+      representer.representable_attrs.size.must_equal 1
+      representer.representable_attrs[:title].must_equal({:representable => true, :as => :track})
+    end
   end
 
 end