+++ CHANGES.md
@@ -1,11 +1,15 @@
 h2. 1.8.0
 
 * Major API change: Remove defaults for collections. This fixes a major design flaw - when parsing a document a collection would be reset to `[]` even if it is not present in the parsed document.
+Definition:
 * Minor API change: Rename `Definition#sought_type` to `#deserialize_class`.
+* Removed #default, #attribute, #content. #from is replaced by #as
 -> constantize :class etc (requires AS)
 -> make all options lambda-able
 -> make major steps lambda-able
 -> strategies for deserialization (lambda-able!)
+
+* Definition#name and #as are _always_ strings.
 
 h2. 1.7.7